Upgraded to node-5.9.0, with assorted fixes
Moving back to `let`s instead of `var`s because what the heck, I control
the versioning. Will probably add a `--harmony` flag to node next
commit.

* Path to tslint is computed via `process.env.NugetMachineInstallRoot`
by default to work with non-standard (read: most larger) build systems.
* `LintRunner`::`lintFolder` returns undefined instead of throwing an
error. The promise chain was stopping because of errors.
* Removed unnecessary `(c) Microsoft ` docs. This is a personal project
so that was inaccurate.
* Pretty TS output without sourceMap